test(eve): parallelize the scenario suite and split the dev-server file
The scenario tier ran all 64 files serially, taking ~8 minutes in CI.
Files now run concurrently across four forked workers against a shared,
run-scoped eve tarball:

- remove the fileParallelism: false constraint and document the
  parallel-safety invariants scenario files must uphold
- stop the framework-next/nuxt scenarios from rebuilding the shared
  workspace dist at test time (the one real parallelism blocker); they
  now build only the framework app against the prebuilt dist
- pack the scenario tarball with lifecycle scripts disabled and require
  a prebuilt dist, instead of paying for prepack to clean and rebuild
  the workspace a second time; the tarball directory is mkdtemp-scoped
  with teardown so concurrent invocations cannot clobber each other
- split the 124s dev-server.scenario.test.ts into four themed files
  (rebuild transactions, live connections, workflow generations, app
  layouts) so the suite's longest pole parallelizes; all 13 tests are
  preserved
- replace the chaos health-probe wall-clock floor with deterministic
  per-round progress checks so the assertion stays stable under load

Local wall time drops from ~311s serial to ~112s.
