+++ apps/frameworks/nuxt/package.json
@@ -4,6 +4,7 @@
   "type": "module",
   "scripts": {
     "build": "pnpm --filter eve build && nuxt build",
+    "build:nuxt": "nuxt build",
     "dev": "nuxt dev",
     "generate": "nuxt generate",
     "preview": "nuxt preview",
